% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FID.R
\name{accessor<-,FID-method}
\alias{accessor<-,FID-method}
\title{accessor<-}
\usage{
\S4method{accessor}{FID}(x) <- value
}
\arguments{
\item{x}{FID object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
