% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\name{accessor<-,FNG-method}
\alias{accessor<-,FNG-method}
\title{accessor<-}
\usage{
\S4method{accessor}{FNG}(x) <- value
}
\arguments{
\item{x}{FNG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
