% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OID.R
\name{accessor<-,OID-method}
\alias{accessor<-,OID-method}
\title{accessor<-}
\usage{
\S4method{accessor}{OID}(x) <- value
}
\arguments{
\item{x}{OID object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
