% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\name{accessor<-,ONG-method}
\alias{accessor<-,ONG-method}
\title{accessor<-}
\usage{
\S4method{accessor}{ONG}(x) <- value
}
\arguments{
\item{x}{ONG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
