% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{accessor<-}
\alias{accessor<-}
\title{accessor<-}
\usage{
accessor(x) <- value
}
\arguments{
\item{x}{the accessor to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for accessor field
}
\keyword{internal}
