% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{analyte<-}
\alias{analyte<-}
\title{analyte<-}
\usage{
analyte(x) <- value
}
\arguments{
\item{x}{the analyte to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for analyte field
}
\keyword{internal}
