% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{analyte_description<-,FOM-method}
\alias{analyte_description<-,FOM-method}
\title{analyte_description<-}
\usage{
\S4method{analyte_description}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
