% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{analyte_description}
\alias{analyte_description}
\title{analyte_description}
\usage{
analyte_description(x)
}
\arguments{
\item{x}{the analyte_description to get}
}
\value{
the value
}
\description{
generic getter for analyte_description field
}
