% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_MAMS.R
\name{convert_AnnData_to_MAMS}
\alias{convert_AnnData_to_MAMS}
\title{Converts a AnnData object to a MAMS object}
\usage{
convert_AnnData_to_MAMS(object_list, observation_subsets, dataset_id)
}
\arguments{
\item{object_list}{A named list of AnnData objects
objects to be converted to MAMS format}

\item{observation_subsets}{A vector with same length as object_list
indicating the observation subset name for each
AnnData object. One of: full, filtered, threshold, detected,
nonartifact, clean.}

\item{dataset_id}{Specify id of the dataset}
}
\value{
A MAMS object containing all the extracted metadata fields.
}
\description{
Converts a AnnData object to a MAMS object
}
\examples{
library(reticulate)
anndata <- import("anndata")
counts <- matrix(rpois((500*200), 1), nrow = 500, ncol = 200, 
    dimnames = list(paste0("Row", 1:500), paste0("Col", 1:200)))
adata <- anndata$AnnData(X = counts)
adata$obsm[["X_pca"]] <- counts[, 1:2]
adata$obsm[["X_tsne"]] <- counts[, 3:4]
mams <- convert_AnnData_to_MAMS(object_list = list(adata = adata), 
    observation_subsets = c("full"), dataset_id = "dataset1")
print(mams)
}
