% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json_converter.R
\name{convert_MAMS_to_list}
\alias{convert_MAMS_to_list}
\title{Convert input MAMS object to a list of list structre
Useful for storing mams object into the uns slot of AnnData object}
\usage{
convert_MAMS_to_list(mams)
}
\arguments{
\item{mams}{MAMS object}
}
\value{
A list of list
}
\description{
Convert input MAMS object to a list of list structre
Useful for storing mams object into the uns slot of AnnData object
}
