% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_MAMS.R
\name{convert_SCE_to_MAMS}
\alias{convert_SCE_to_MAMS}
\title{Converts a SingleCellExperiment object to a MAMS object}
\usage{
convert_SCE_to_MAMS(object_list, observation_subsets, dataset_id)
}
\arguments{
\item{object_list}{A named list of SingleCellExperiment
objects to be converted to MAMS format}

\item{observation_subsets}{A vector with same length as object_list
indicating the observation subset name for each
SingleCellExperiment object. One of: full, filtered, threshold, detected,
nonartifact, clean.}

\item{dataset_id}{Specify id of the dataset}
}
\value{
A MAMS object containing all the extracted metadata fields.
}
\description{
Converts a SingleCellExperiment object to a MAMS object
}
\examples{
\dontrun{
library(SingleCellExperiment)
counts <- matrix(rpois((500*200), 1), nrow = 500, ncol = 200, 
    dimnames = list(paste0("Row", 1:500), paste0("Col", 1:200)))
sce <- SingleCellExperiment(list(counts=counts))
sce <- scuttle::logNormCounts(sce)
subset_sce <- sce[, 1:100]
mams <- convert_SCE_to_MAMS(object_list = list(sce = sce, 
    subset_sce = subset_sce), observation_subsets = c("full", "subset"), 
    dataset_id = "dataset1")
print(mams)
}
}
