% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_MAMS.R
\name{convert_seurat_to_MAMS}
\alias{convert_seurat_to_MAMS}
\title{Converts Seurat object to a MAMS object}
\usage{
convert_seurat_to_MAMS(object_list, observation_subsets)
}
\arguments{
\item{object_list}{}
}
\description{
Converts Seurat object to a MAMS object
}
