% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\name{create_FEA_object}
\alias{create_FEA_object}
\title{Constructor for the FEA S4 object}
\usage{
create_FEA_object(
  id = NA_character_,
  dataset_id = NA_character_,
  filepath = NA_character_,
  accessor = NA_character_,
  feature_modality = NA_character_,
  reference_database = NA_character_,
  reference_organism = NA_character_,
  record_id = NA_character_
)
}
\arguments{
\item{id}{Parent ID}

\item{dataset_id}{Parent dataset ID}

\item{filepath}{Path to the data file}

\item{accessor}{Accessors used}

\item{feature_modality}{Modality of the feature}

\item{reference_database}{Reference database used}

\item{reference_organism}{Reference organism used}

\item{record_id}{Record ID of object}
}
\value{
a FEA S4 object for use with MAMS
}
\description{
Creates the FEA (feature) object and populates its subfields
}
