% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FID.R
\name{create_FID_object}
\alias{create_FID_object}
\title{Constructor for the FID S4 object}
\usage{
create_FID_object(
  id = NA_character_,
  dataset_id = NA_character_,
  filepath = NA_character_,
  accessor = NA_character_
)
}
\arguments{
\item{id}{Parent ID}

\item{dataset_id}{Parent dataset ID}

\item{filepath}{Path to the data file}

\item{accessor}{Accessors used}
}
\value{
a FID S4 object for use with MAMS
}
\description{
Creates the FID object and populates its subfields
}
