% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{create_FOM_object}
\alias{create_FOM_object}
\title{Constructor function to create a FOM object}
\usage{
create_FOM_object(
  id = NA_character_,
  dataset_id = NA_character_,
  filepath = NA_character_,
  accessor = NA_character_,
  data_type = NA_character_,
  representation = NA_character_,
  representation_description = NA_character_,
  obs_unit = NA_character_,
  processing = NA_character_,
  processing_description = NA_character_,
  analyte = NA_character_,
  analyte_description = NA_character_,
  modality = NA_character_,
  obs_subset = NA_character_,
  obs_subset_description = NA_character_,
  feature_subset = NA_character_,
  feature_subset_description = NA_character_,
  record_id = NA_character_,
  parent_id = NA_character_,
  parent_relationship = NA_character_,
  parent_relationship_description = NA_character_,
  oid = NA_character_,
  fid = NA_character_,
  obs = NA_character_,
  fea = NA_character_,
  ong = NA_character_,
  fng = NA_character_
)
}
\arguments{
\item{id}{}

\item{dataset_id}{}

\item{data_type}{}

\item{representation}{}

\item{representation_description}{}

\item{obs_unit}{}

\item{processing}{}

\item{processing_description}{}

\item{analyte}{}

\item{analyte_description}{}

\item{modality}{}

\item{obs_subset}{}

\item{obs_subset_description}{}

\item{feature_subset}{}

\item{feature_subset_description}{}

\item{record_id}{}

\item{parent_id}{}

\item{parent_relationship}{}

\item{parent_relationship_description}{}

\item{fng}{}
}
\description{
Constructor function to create a FOM object
}
