% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json_converter.R
\name{create_MAMS_from_JSON}
\alias{create_MAMS_from_JSON}
\title{Create a MAMS object from a JSON file}
\usage{
create_MAMS_from_JSON(JSON)
}
\arguments{
\item{JSON}{a JSON file containing the MAMS data}
}
\value{
a MAMS object
}
\description{
Create a MAMS object from a JSON file
}
