% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OBS.R
\name{create_OBS_object}
\alias{create_OBS_object}
\title{Constructor for the OBS (observations) S4 object}
\usage{
create_OBS_object(
  id = NA_character_,
  dataset_id = NA_character_,
  filepath = NA_character_,
  accessor = NA_character_,
  record_id = NA_character_
)
}
\arguments{
\item{id}{Main ID of MAMS object}

\item{dataset_id}{Parent dataset ID}

\item{filepath}{Path to the data file}

\item{accessor}{Accessor}

\item{record_id}{Record ID}
}
\value{
a OBS S4 object for use with MAMS
}
\description{
Creates the FID object and populates its subfields
}
