% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OID.R
\name{create_OID_object}
\alias{create_OID_object}
\title{Constructor for the OID S4 object}
\usage{
create_OID_object(
  id = NA_character_,
  dataset_id = NA_character_,
  filepath = NA_character_,
  accessor = NA_character_
)
}
\arguments{
\item{id}{character}

\item{dataset_id}{character}

\item{filepath}{character}

\item{accessor}{character}
}
\value{
a OID S4 object for use with MAMS
}
\description{
Creates the OID object and populates its subfields
}
