% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{create_REC_object}
\alias{create_REC_object}
\title{Constructor for the REC (provenance records) S4 object}
\usage{
create_REC_object(
  id = NA_character_,
  dataset_id = NA_character_,
  record_id = NA_character_,
  record_package_name = NA_character_,
  record_package_version = NA_character_,
  record_function_name = NA_character_,
  record_function_parameters = NA_character_,
  record_workflow_link = NA_character_,
  record_runtime_start = NA_character_,
  record_runtime_end = NA_character_,
  record_runtime_duration = NA_character_
)
}
\arguments{
\item{id}{Main ID of MAMS object}

\item{dataset_id}{Parent dataset ID}

\item{record_id}{Record ID}

\item{record_package_name}{Name of toolkit}

\item{record_package_version}{Version of toolkit}

\item{record_function_name}{Name of function(s) used}

\item{record_function_parameters}{Parameters passed to function(s)}

\item{record_workflow_link}{Link to the workflow of the toolkit}

\item{record_runtime_start}{Runtime start}

\item{record_runtime_end}{Runtime end}

\item{record_runtime_duration}{Run duration}
}
\value{
a REC S4 object for use with MAMS
}
\description{
Creates the FID object and populates its subfields
}
