% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{data_type<-}
\alias{data_type<-}
\title{data_type<-}
\usage{
data_type(x) <- value
}
\arguments{
\item{x}{the data_type to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for data_type field
}
\keyword{internal}
