% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\name{dataset_id<-,FEA-method}
\alias{dataset_id<-,FEA-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{FEA}(x) <- value
}
\arguments{
\item{x}{FEA object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
