% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FID.R
\name{dataset_id<-,FID-method}
\alias{dataset_id<-,FID-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{FID}(x) <- value
}
\arguments{
\item{x}{FID object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
