% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\name{dataset_id<-,FNG-method}
\alias{dataset_id<-,FNG-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{FNG}(x) <- value
}
\arguments{
\item{x}{FNG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
