% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OID.R
\name{dataset_id<-,OID-method}
\alias{dataset_id<-,OID-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{OID}(x) <- value
}
\arguments{
\item{x}{OID object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
