% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\name{dataset_id<-,ONG-method}
\alias{dataset_id<-,ONG-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{ONG}(x) <- value
}
\arguments{
\item{x}{ONG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
