% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{dataset_id<-,REC-method}
\alias{dataset_id<-,REC-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{REC}(x) <- value
}
\arguments{
\item{x}{REC object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
