% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{dataset_id<-,FOM-method}
\alias{dataset_id<-,FOM-method}
\title{dataset_id<-}
\usage{
\S4method{dataset_id}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\description{
setter
}
