% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{dataset_id<-}
\alias{dataset_id<-}
\title{dataset_id<-}
\usage{
dataset_id(x) <- value
}
\arguments{
\item{x}{the dataset_id to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for dataset_id field
}
