% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\name{edge_metric<-,FNG-method}
\alias{edge_metric<-,FNG-method}
\title{edge_metric<-}
\usage{
\S4method{edge_metric}{FNG}(x) <- value
}
\arguments{
\item{x}{FNG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
