% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\name{edge_metric<-,ONG-method}
\alias{edge_metric<-,ONG-method}
\title{edge_metric<-}
\usage{
\S4method{edge_metric}{ONG}(x) <- value
}
\arguments{
\item{x}{ONG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
