% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{edge_metric<-}
\alias{edge_metric<-}
\title{edge_metric<-}
\usage{
edge_metric(x) <- value
}
\arguments{
\item{x}{the edge_metric to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for edge_metric field
}
\keyword{internal}
