% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\docType{class}
\name{FEA-class}
\alias{FEA-class}
\title{Class to create a FEA object}
\description{
Class to create a FEA object
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{dataset_id}}{character.}

\item{\code{filepath}}{character.}

\item{\code{accessor}}{character.}

\item{\code{feature_modality}}{character.}

\item{\code{reference_database}}{character.}

\item{\code{reference_organism}}{character.}

\item{\code{record_id}}{character.}
}}

