% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{fea}
\alias{fea}
\title{fea}
\usage{
fea(mams, fea_id, key)
}
\arguments{
\item{mams}{the MAMS object}

\item{fea_id}{the FEA ID}

\item{key}{key of the value wanted}
}
\value{
the value
}
\description{
MAMS generic getter for FEA objects
}
