% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{fea,MAMS,character,character-method}
\alias{fea,MAMS,character,character-method}
\title{fea}
\usage{
\S4method{fea}{MAMS,character,character}(mams, fea_id, key)
}
\arguments{
\item{mams}{MAMS object}

\item{fea_id}{FEA subobject ID}

\item{key}{key}
}
\value{
value of key
}
\description{
FEA getter
}
