% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\name{feature_modality,FEA-method}
\alias{feature_modality,FEA-method}
\title{feature_modality}
\usage{
\S4method{feature_modality}{FEA}(x)
}
\arguments{
\item{x}{FEA object}
}
\value{
the value
}
\description{
getter
}
