% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{feature_modality}
\alias{feature_modality}
\title{feature_modality}
\usage{
feature_modality(x)
}
\arguments{
\item{x}{the feature_modality to get}
}
\value{
the value
}
\description{
generic getter for feature_modality field
}
\keyword{internal}
