% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{feature_modality<-}
\alias{feature_modality<-}
\title{feature_modality<-}
\usage{
feature_modality(x) <- value
}
\arguments{
\item{x}{the feature_modality to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for feature_modality field
}
\keyword{internal}
