% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{feature_subset<-,FOM-method}
\alias{feature_subset<-,FOM-method}
\title{feature_subset<-}
\usage{
\S4method{feature_subset}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
