% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{feature_subset<-}
\alias{feature_subset<-}
\title{feature_subset<-}
\usage{
feature_subset(x) <- value
}
\arguments{
\item{x}{the feature_subset to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for feature_subset field
}
