% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{feature_subset_description,FOM-method}
\alias{feature_subset_description,FOM-method}
\title{feature_subset_description}
\usage{
\S4method{feature_subset_description}{FOM}(x)
}
\arguments{
\item{x}{FOM object}
}
\description{
getter
}
