% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{feature_subset_description}
\alias{feature_subset_description}
\title{feature_subset_description}
\usage{
feature_subset_description(x)
}
\arguments{
\item{x}{the obs_subset_descrifeature_subset_descriptionption to get}
}
\value{
the value
}
\description{
generic getter for feature_subset_description field
}
\keyword{internal}
