% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{feature_subset_description<-}
\alias{feature_subset_description<-}
\title{feature_subset_description<-}
\usage{
feature_subset_description(x) <- value
}
\arguments{
\item{x}{the feature_subset_description to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for feature_subset_description field
}
