% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FID.R
\docType{class}
\name{FID-class}
\alias{FID-class}
\title{class FID}
\value{
a FEA class for use with MAMS
}
\description{
Stores feature ID class
}
\details{
Define the FID (feature ID) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character}

\item{\code{dataset_id}}{character}

\item{\code{filepath}}{character}

\item{\code{accessor}}{character}
}}

