% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{fid}
\alias{fid}
\title{fid}
\usage{
fid(mams, fid_id, key)
}
\arguments{
\item{mams}{the MAMS object}

\item{fid_id}{the FID ID}

\item{key}{key of the value wanted}
}
\value{
the value
}
\description{
MAMS generic getter for FID objects
}
