% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{fid<-,MAMS,character,character-method}
\alias{fid<-,MAMS,character,character-method}
\title{fid<-}
\usage{
\S4method{fid}{MAMS,character,character}(mams, fid_id, key) <- value
}
\arguments{
\item{mams}{MAMS object}

\item{fid_id}{FID subobject ID}

\item{key}{key}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
FID setter
}
