% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{fid<-}
\alias{fid<-}
\title{fid<-}
\usage{
fid(mams, fid_id, key) <- value
}
\arguments{
\item{mams}{the MAMS object}

\item{fid_id}{the FID ID}

\item{key}{key of the value wanted}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
MAMS generic setter for FID objects
}
