% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FID.R
\name{filepath<-,FID-method}
\alias{filepath<-,FID-method}
\title{filepath<-}
\usage{
\S4method{filepath}{FID}(x) <- value
}
\arguments{
\item{x}{FID object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
