% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\name{filepath<-,FNG-method}
\alias{filepath<-,FNG-method}
\title{filepath<-}
\usage{
\S4method{filepath}{FNG}(x) <- value
}
\arguments{
\item{x}{FNG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
