% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{filepath<-,FOM-method}
\alias{filepath<-,FOM-method}
\title{filepath<-}
\usage{
\S4method{filepath}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
