% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OBS.R
\name{filepath<-,OBS-method}
\alias{filepath<-,OBS-method}
\title{filepath<-}
\usage{
\S4method{filepath}{OBS}(x) <- value
}
\arguments{
\item{x}{OBS object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
