% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\name{filepath<-,ONG-method}
\alias{filepath<-,ONG-method}
\title{filepath<-}
\usage{
\S4method{filepath}{ONG}(x) <- value
}
\arguments{
\item{x}{ONG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
