% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{filepath<-}
\alias{filepath<-}
\title{filepath<-}
\usage{
filepath(x) <- value
}
\arguments{
\item{x}{the filepath to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for filepath field
}
\keyword{internal}
