% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\docType{class}
\name{FNG-class}
\alias{FNG-class}
\title{class FNG}
\value{
a FNG S4 object for use with MAMS
}
\description{
Stores feature neighborhood graph data
}
\details{
Define the FNG (feature neighborhood graph) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character}

\item{\code{dataset_id}}{character}

\item{\code{filepath}}{character}

\item{\code{accessor}}{character}

\item{\code{parent_id}}{character}

\item{\code{record_id}}{character}

\item{\code{edge_metric}}{character}

\item{\code{metric_type}}{character}
}}

