% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{fng,MAMS,character,character-method}
\alias{fng,MAMS,character,character-method}
\alias{fng<-,MAMS,character,character-method}
\title{fng}
\usage{
\S4method{fng}{MAMS,character,character}(mams, fng_id, key)

\S4method{fng}{MAMS,character,character}(mams, fng_id, key) <- value
}
\arguments{
\item{mams}{MAMS object}

\item{fng_id}{FNG subobject ID}

\item{key}{key}

\item{value}{value}
}
\value{
value of key

nothing (setter)
}
\description{
FNG getter

FNG setter
}
