% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{fng<-}
\alias{fng<-}
\title{fng<-}
\usage{
fng(mams, fng_id, key) <- value
}
\arguments{
\item{mams}{the MAMS object}

\item{fng_id}{the FNG ID}

\item{key}{key of the value wanted}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
MAMS generic setter for FNG objects
}
