% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\docType{class}
\name{FOM-class}
\alias{FOM-class}
\title{class FOM}
\value{
a FOM class for further use with other MAMS objects and sub-objects
}
\description{
Stores the main feature observation matrix
}
\details{
Define the Feature Observation Matrix (FOM) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{dataset_id}}{character.}

\item{\code{filepath}}{character.}

\item{\code{accessor}}{character.}

\item{\code{data_type}}{character.}

\item{\code{representation}}{character.}

\item{\code{representation_description}}{character.}

\item{\code{obs_unit}}{character.}

\item{\code{obs_unit_description}}{character.}

\item{\code{processing}}{character.}

\item{\code{processing_description}}{character.}

\item{\code{analyte}}{List or character.}

\item{\code{analyte_description}}{List or character.}

\item{\code{modality}}{List or character.}

\item{\code{obs_subset}}{character.}

\item{\code{obs_subset_description}}{character.}

\item{\code{feature_subset}}{character.}

\item{\code{feature_subset_description}}{character.}

\item{\code{record_id}}{character.}

\item{\code{parent_id}}{character or list.}

\item{\code{parent_relationship}}{character.}

\item{\code{parent_relationship_description}}{character.}

\item{\code{oid}}{character.}

\item{\code{fid}}{character.}

\item{\code{obs}}{character.}

\item{\code{fea}}{character.}

\item{\code{ong}}{character.}

\item{\code{fng}}{character.}
}}

