% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{fom}
\alias{fom}
\title{fom}
\usage{
fom(mams, fom_id, key)
}
\arguments{
\item{mams}{the MAMS object}

\item{fom_id}{the FOM ID}

\item{key}{key of the value wanted}
}
\value{
the value
}
\description{
MAMS generic getter for FOM objects
}
