% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{fom,MAMS,character,character-method}
\alias{fom,MAMS,character,character-method}
\title{fom}
\usage{
\S4method{fom}{MAMS,character,character}(mams, fom_id, key)
}
\arguments{
\item{mams}{MAMS object}

\item{fom_id}{FOM subobject ID}

\item{key}{key}
}
\value{
value of key
}
\description{
FOM getter
}
