% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\docType{class}
\name{MAMS-class}
\alias{MAMS-class}
\title{class MAMS}
\value{
the main MAMS class
}
\description{
Stores the MAMS data as a list of lists (each of S4 objects).
}
\details{
Define the main MAMS (Matrix and Analysis Metadata Standards) S4 object, used to store all of the other objects.
}
\section{Slots}{

\describe{
\item{\code{FOM}}{list}

\item{\code{FEA}}{list}

\item{\code{OBS}}{list}

\item{\code{FID}}{list}

\item{\code{OID}}{list}

\item{\code{REC}}{list}

\item{\code{ONG}}{list}

\item{\code{FNG}}{list}
}}

