% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\name{metric_type<-,FNG-method}
\alias{metric_type<-,FNG-method}
\title{metric_type<-}
\usage{
\S4method{metric_type}{FNG}(x) <- value
}
\arguments{
\item{x}{FNG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
