% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\name{metric_type<-,ONG-method}
\alias{metric_type<-,ONG-method}
\title{metric_type<-}
\usage{
\S4method{metric_type}{ONG}(x) <- value
}
\arguments{
\item{x}{ONG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
