% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{modality}
\alias{modality}
\title{processing_modalitydescription}
\usage{
modality(x)
}
\arguments{
\item{x}{the modality to get}
}
\value{
the value
}
\description{
generic getter for modality field
}
