% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{modality<-}
\alias{modality<-}
\title{modality<-}
\usage{
modality(x) <- value
}
\arguments{
\item{x}{the modality to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for modality field
}
