% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OBS.R
\docType{class}
\name{OBS-class}
\alias{OBS-class}
\title{Class to create a OBS object}
\description{
Class to create a OBS object
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{dataset_id}}{character.}

\item{\code{filepath}}{character.}

\item{\code{accessor}}{character.}

\item{\code{record_id}}{character.}
}}

