% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{obs}
\alias{obs}
\title{obs}
\usage{
obs(mams, obs_id, key)
}
\arguments{
\item{mams}{the MAMS object}

\item{obs_id}{the OBS ID}

\item{key}{key of the value wanted}
}
\value{
the value
}
\description{
MAMS generic getter for OBS objects
}
