% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{obs<-,MAMS,character,character-method}
\alias{obs<-,MAMS,character,character-method}
\title{obs<-}
\usage{
\S4method{obs}{MAMS,character,character}(mams, obs_id, key) <- value
}
\arguments{
\item{mams}{MAMS object}

\item{obs_id}{OBS subobject ID}

\item{key}{key}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
OBS setter
}
