% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{obs<-}
\alias{obs<-}
\title{obs<-}
\usage{
obs(mams, obs_id, key) <- value
}
\arguments{
\item{mams}{the MAMS object}

\item{obs_id}{the OBS ID}

\item{key}{key of the value wanted}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
MAMS generic setter for OBS objects
}
