% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{obs_subset_description,FOM-method}
\alias{obs_subset_description,FOM-method}
\title{obs_subset_description}
\usage{
\S4method{obs_subset_description}{FOM}(x)
}
\arguments{
\item{x}{FOM object}
}
\value{
the value
}
\description{
getter
}
