% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{obs_subset}
\alias{obs_subset}
\title{obs_subset}
\usage{
obs_subset(x)
}
\arguments{
\item{x}{the obs_subset to get}
}
\value{
the value
}
\description{
generic getter for obs_subset field
}
\keyword{internal}
