% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{obs_subset<-,FOM-method}
\alias{obs_subset<-,FOM-method}
\title{obs_subset<-}
\usage{
\S4method{obs_subset}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\description{
setter
}
