% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{obs_subset_description<-}
\alias{obs_subset_description<-}
\title{obs_subset_description<-}
\usage{
obs_subset_description(x) <- value
}
\arguments{
\item{x}{the obs_subset_description to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for obs_subset_description field
}
