% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{obs_unit<-}
\alias{obs_unit<-}
\title{obs_unit<-}
\usage{
obs_unit(x) <- value
}
\arguments{
\item{x}{the obs_unit to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for obs_unit field
}
\keyword{internal}
