% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{obs_unit_description,FOM-method}
\alias{obs_unit_description,FOM-method}
\alias{obs_unit_description<-,FOM-method}
\title{obs_unit_description}
\usage{
\S4method{obs_unit_description}{FOM}(x)

\S4method{obs_unit_description}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\value{
nothing (setter)

nothing (setter)
}
\description{
setter

setter
}
