% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OID.R
\docType{class}
\name{OID-class}
\alias{OID-class}
\title{class OID}
\value{
the OID class
}
\description{
Stores Observation ID class
}
\details{
Define the OID (observation ID) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character}

\item{\code{dataset_id}}{character}

\item{\code{filepath}}{character}

\item{\code{accessor}}{character}
}}

