% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{oid}
\alias{oid}
\title{oid}
\usage{
oid(mams, oid_id, key)
}
\arguments{
\item{mams}{the MAMS object}

\item{oid_id}{the OID ID}

\item{key}{key of the value wanted}
}
\value{
the value
}
\description{
MAMS generic getter for OID objects
}
