% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{oid<-,MAMS,character,character-method}
\alias{oid<-,MAMS,character,character-method}
\title{oid<-}
\usage{
\S4method{oid}{MAMS,character,character}(mams, oid_id, key) <- value
}
\arguments{
\item{mams}{MAMS object}

\item{oid_id}{OID subobject ID}

\item{key}{key}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
OID setter
}
