% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{oid<-}
\alias{oid<-}
\title{oid<-}
\usage{
oid(mams, oid_id, key) <- value
}
\arguments{
\item{mams}{the MAMS object}

\item{oid_id}{the OID ID}

\item{key}{key of the value wanted}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
MAMS generic setter for OID objects
}
