% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\docType{class}
\name{ONG-class}
\alias{ONG-class}
\title{class ONG}
\value{
the ONG class for use with MAMS
}
\description{
Stores observation neighborhood graph data
}
\details{
Define the ONG (observation neighborhood graph) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character}

\item{\code{dataset_id}}{character}

\item{\code{filepath}}{character}

\item{\code{accessor}}{character}

\item{\code{parent_id}}{character}

\item{\code{record_id}}{character}

\item{\code{edge_metric}}{character}

\item{\code{metric_type}}{character}
}}

