% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{ong,MAMS,character,character-method}
\alias{ong,MAMS,character,character-method}
\title{ong}
\usage{
\S4method{ong}{MAMS,character,character}(mams, ong_id, key)
}
\arguments{
\item{mams}{MAMS object}

\item{ong_id}{ONG subobject ID}

\item{key}{key}
}
\value{
value of key
}
\description{
ONG getter
}
