% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{ong<-}
\alias{ong<-}
\title{ong<-}
\usage{
ong(mams, ong_id, key) <- value
}
\arguments{
\item{mams}{the MAMS object}

\item{ong_id}{the ONG ID}

\item{key}{key of the value wanted}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
MAMS generic setter for ONG objects
}
