% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{parent_id}
\alias{parent_id}
\title{parent_id}
\usage{
parent_id(x)
}
\arguments{
\item{x}{the parent_id to get}
}
\value{
the value
}
\description{
generic getter for parent_id field
}
\keyword{internal}
