% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_ONG.R
\name{parent_id<-,ONG-method}
\alias{parent_id<-,ONG-method}
\title{parent_id<-}
\usage{
\S4method{parent_id}{ONG}(x) <- value
}
\arguments{
\item{x}{ONG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
