% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{parent_relationship<-,FOM-method}
\alias{parent_relationship<-,FOM-method}
\title{parent_relationship<-}
\usage{
\S4method{parent_relationship}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
