% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{parent_relationship}
\alias{parent_relationship}
\title{parent_relationship}
\usage{
parent_relationship(x)
}
\arguments{
\item{x}{the parent_relationship to get}
}
\value{
the value
}
\description{
generic getter for parent_relationship field
}
\keyword{internal}
