% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{parent_relationship_description,FOM-method}
\alias{parent_relationship_description,FOM-method}
\title{parent_relationship_description}
\usage{
\S4method{parent_relationship_description}{FOM}(x)
}
\arguments{
\item{x}{FOM object}
}
\value{
the value
}
\description{
getter
}
