% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{parent_relationship_description}
\alias{parent_relationship_description}
\title{parent_relationship_description}
\usage{
parent_relationship_description(x)
}
\arguments{
\item{x}{the parent_relationship_description to get}
}
\value{
the value
}
\description{
generic getter for parent_relationship_description field
}
\keyword{internal}
