% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{parent_relationship_description<-}
\alias{parent_relationship_description<-}
\title{parent_relationship_description<-}
\usage{
parent_relationship_description(x) <- value
}
\arguments{
\item{x}{the parent_relationship_description to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for parent_relationship_description field
}
\keyword{internal}
