% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{obs_unit_description<-,FOM-method}
\alias{obs_unit_description<-,FOM-method}
\alias{processing,FOM-method}
\title{processing}
\usage{
\S4method{obs_unit_description}{FOM}(x) <- value

\S4method{processing}{FOM}(x)
}
\arguments{
\item{x}{FOM object}
}
\value{
the value

the value
}
\description{
getter

getter
}
