% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{processing<-}
\alias{processing<-}
\title{processing<-}
\usage{
processing(x) <- value
}
\arguments{
\item{x}{the processing to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for processing field
}
\keyword{internal}
