% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{processing_description}
\alias{processing_description}
\title{processing_description}
\usage{
processing_description(x)
}
\arguments{
\item{x}{the processing_description to get}
}
\value{
the value
}
\description{
generic getter for processing_description field
}
\keyword{internal}
