% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{processing_description<-,FOM-method}
\alias{processing_description<-,FOM-method}
\title{processing_description<-}
\usage{
\S4method{processing_description}{FOM}(x) <- value
}
\arguments{
\item{x}{FOM object}

\item{value}{value}
}
\description{
setter
}
