% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\docType{class}
\name{REC-class}
\alias{REC-class}
\title{class REC}
\description{
Stores provenance about the dataset and also other objects
}
\section{Slots}{

\describe{
\item{\code{id}}{}

\item{\code{dataset_id}}{}

\item{\code{record_id}}{}

\item{\code{record_package_name}}{}

\item{\code{record_package_version}}{}

\item{\code{record_function_name}}{}

\item{\code{record_function_parameters}}{}

\item{\code{record_workflow_link}}{}

\item{\code{record_runtime_start}}{}

\item{\code{record_runtime_end}}{}

\item{\code{record_runtime_duration}}{}
}}

