% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{rec}
\alias{rec}
\title{rec}
\usage{
rec(mams, rec_id, key)
}
\arguments{
\item{mams}{the MAMS object}

\item{rec_id}{the REC ID}

\item{key}{key of the value wanted}
}
\value{
the value
}
\description{
MAMS generic getter for REC objects
}
