% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{rec,MAMS,character,character-method}
\alias{rec,MAMS,character,character-method}
\title{rec}
\usage{
\S4method{rec}{MAMS,character,character}(mams, rec_id, key)
}
\arguments{
\item{mams}{MAMS object}

\item{rec_id}{REC subobject ID}

\item{key}{key}
}
\value{
value of key
}
\description{
REC getter
}
