% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{rec<-}
\alias{rec<-}
\title{rec<-}
\usage{
rec(mams, rec_id, key) <- value
}
\arguments{
\item{mams}{the MAMS object}

\item{rec_id}{the REC ID}

\item{key}{key of the value wanted}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
MAMS generic setter for REC objects
}
