% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{record_function_name,REC-method}
\alias{record_function_name,REC-method}
\title{record_function_name}
\usage{
\S4method{record_function_name}{REC}(x)
}
\arguments{
\item{x}{REC object}
}
\value{
value
}
\description{
getter
}
