% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_function_name<-}
\alias{record_function_name<-}
\title{record_function_name<-}
\usage{
record_function_name(x) <- value
}
\arguments{
\item{x}{the record_function_name to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for record_function_name field
}
