% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_function_parameters}
\alias{record_function_parameters}
\title{record_function_parameters}
\usage{
record_function_parameters(x)
}
\arguments{
\item{x}{the record_function_parameters to get}
}
\value{
the value
}
\description{
generic getter for record_function_parameters field
}
\keyword{internal}
