% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FNG.R
\name{record_id<-,FNG-method}
\alias{record_id<-,FNG-method}
\title{record_id<-}
\usage{
\S4method{record_id}{FNG}(x) <- value
}
\arguments{
\item{x}{FNG object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
