% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_id}
\alias{record_id}
\title{record_id}
\usage{
record_id(x)

record_id(x)
}
\arguments{
\item{x}{the record_id to get}
}
\value{
the value

the value
}
\description{
generic getter for record_id field

generic getter for record_id field
}
