% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OBS.R
\name{record_id<-,OBS-method}
\alias{record_id<-,OBS-method}
\title{record_id<-}
\usage{
\S4method{record_id}{OBS}(x) <- value
}
\arguments{
\item{x}{OBS object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
