% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_id<-}
\alias{record_id<-}
\title{record_id<-}
\usage{
record_id(x) <- value

record_id(x) <- value
}
\arguments{
\item{x}{the record_id to get}

\item{value}{the new value}
}
\value{
nothing (setter)

nothing (setter)
}
\description{
generic getter for record_id field

generic getter for record_id field
}
\keyword{internal}
