% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{record_package_name<-,REC-method}
\alias{record_package_name<-,REC-method}
\title{record_package_name<-}
\usage{
\S4method{record_package_name}{REC}(x) <- value
}
\arguments{
\item{x}{REC object}

\item{value}{value}
}
\value{
nothing (setter)
}
\description{
setter
}
