% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_package_name<-}
\alias{record_package_name<-}
\title{record_package_name<-}
\usage{
record_package_name(x) <- value
}
\arguments{
\item{x}{the record_package_name to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for record_package_name field
}
\keyword{internal}
