% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{record_package_version,REC-method}
\alias{record_package_version,REC-method}
\title{record_package_version}
\usage{
\S4method{record_package_version}{REC}(x)
}
\arguments{
\item{x}{REC object}
}
\value{
value
}
\description{
getter
}
