% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_package_version<-}
\alias{record_package_version<-}
\title{record_package_version<-}
\usage{
record_package_version(x) <- value
}
\arguments{
\item{x}{the record_package_version to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for record_package_version field
}
