% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_runtime_duration}
\alias{record_runtime_duration}
\title{record_runtime_duration}
\usage{
record_runtime_duration(x)
}
\arguments{
\item{x}{the record_runtime_duration to get}
}
\value{
the value
}
\description{
generic getter for record_runtime_duration field
}
