% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{record_runtime_duration,REC-method}
\alias{record_runtime_duration,REC-method}
\title{record_runtime_duration}
\usage{
\S4method{record_runtime_duration}{REC}(x)
}
\arguments{
\item{x}{REC object}
}
\value{
value
}
\description{
getter
}
