% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_runtime_end}
\alias{record_runtime_end}
\title{record_runtime_end}
\usage{
record_runtime_end(x)
}
\arguments{
\item{x}{the record_runtime_end to get}
}
\value{
the value
}
\description{
generic getter for record_runtime_end field
}
