% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_runtime_start}
\alias{record_runtime_start}
\title{record_runtime_start}
\usage{
record_runtime_start(x)
}
\arguments{
\item{x}{the record_runtime_start to get}
}
\value{
the value
}
\description{
generic getter for record_runtime_start field
}
\keyword{internal}
