% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{record_runtime_start,REC-method}
\alias{record_runtime_start,REC-method}
\title{record_runtime_start}
\usage{
\S4method{record_runtime_start}{REC}(x)
}
\arguments{
\item{x}{REC object}
}
\value{
value
}
\description{
getter
}
