% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_REC.R
\name{record_workflow_link,REC-method}
\alias{record_workflow_link,REC-method}
\title{record_workflow_link}
\usage{
\S4method{record_workflow_link}{REC}(x)
}
\arguments{
\item{x}{REC object}
}
\value{
value
}
\description{
getter
}
