% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{record_workflow_link<-}
\alias{record_workflow_link<-}
\title{record_workflow_link<-}
\usage{
record_workflow_link(x) <- value
}
\arguments{
\item{x}{the record_workflow_link to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for record_workflow_link field
}
