% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\name{reference_database,FEA-method}
\alias{reference_database,FEA-method}
\title{reference_database}
\usage{
\S4method{reference_database}{FEA}(x)
}
\arguments{
\item{x}{FEA object}
}
\value{
the value
}
\description{
getter
}
