% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\name{reference_database<-,FEA-method}
\alias{reference_database<-,FEA-method}
\title{reference_database<-}
\usage{
\S4method{reference_database}{FEA}(x) <- value
}
\arguments{
\item{x}{FEA object}

\item{value}{value}
}
\description{
setter
}
