% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{reference_database}
\alias{reference_database}
\title{reference_database}
\usage{
reference_database(x)
}
\arguments{
\item{x}{the reference_database to get}
}
\value{
the value
}
\description{
generic getter for reference_database field
}
\keyword{internal}
