% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\name{reference_organism,FEA-method}
\alias{reference_organism,FEA-method}
\title{reference_organism}
\usage{
\S4method{reference_organism}{FEA}(x)
}
\arguments{
\item{x}{FEA object}
}
\value{
the value

nothing (setter)
}
\description{
getter
}
