% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{reference_organism<-}
\alias{reference_organism<-}
\title{reference_organism<-}
\usage{
reference_organism(x) <- value
}
\arguments{
\item{x}{the reference_organism to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for reference_organism field
}
