% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{representation,FOM-method}
\alias{representation,FOM-method}
\title{representation}
\usage{
\S4method{representation}{FOM}(x)
}
\arguments{
\item{x}{FOM object}
}
\value{
the value
}
\description{
getter
}
