% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{representation}
\alias{representation}
\title{representation}
\usage{
representation(x)
}
\arguments{
\item{x}{the representation to get}
}
\value{
the value
}
\description{
generic getter for representation field
}
