% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{representation<-}
\alias{representation<-}
\title{representation<-}
\usage{
representation(x) <- value
}
\arguments{
\item{x}{the representation to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for representation field
}
