% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{representation_description,FOM-method}
\alias{representation_description,FOM-method}
\title{representation_description}
\usage{
\S4method{representation_description}{FOM}(x)
}
\arguments{
\item{x}{FOM object}
}
\description{
getter
}
