% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{representation_description}
\alias{representation_description}
\title{representation_description}
\usage{
representation_description(x)
}
\arguments{
\item{x}{the representation_description to get}
}
\value{
the value
}
\description{
generic getter for representation_description field
}
