% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_generics.R
\name{representation_description<-}
\alias{representation_description<-}
\title{representation_description<-}
\usage{
representation_description(x) <- value
}
\arguments{
\item{x}{the representation_description to get}

\item{value}{the new value}
}
\value{
nothing (setter)
}
\description{
generic getter for representation_description field
}
