% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json_converter.R
\name{write_MAMS_to_JSON}
\alias{write_MAMS_to_JSON}
\title{Create a JSON object from MAMS}
\usage{
write_MAMS_to_JSON(MAMS, filepath)
}
\arguments{
\item{MAMS}{a mams object}
}
\value{
a JSON object
}
\description{
Create a JSON object from MAMS
}
