% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Json_converter.R
\name{write_MAMS_to_output}
\alias{write_MAMS_to_output}
\title{Create a JSON object from MAMS}
\usage{
write_MAMS_to_output(MAMS, filepath, format)
}
\arguments{
\item{MAMS}{a mams object}

\item{filepath}{specify an output filepath}

\item{format}{specify if the output needed is YAML or JSON}
}
\value{
a JSON object
}
\description{
Create a JSON object from MAMS
}
