% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_MAMS.R
\name{check_MAMS}
\alias{check_MAMS}
\title{Validate mams object}
\usage{
check_MAMS(mams_object)
}
\arguments{
\item{mams_object}{Input MAMS object to be checked}
}
\description{
Internal check function to determine if all necessary MAMS object slots
exist within the object
}
