% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_MAMS.R
\name{convert_seurat_to_MAMS}
\alias{convert_seurat_to_MAMS}
\title{Converts Seurat object to a MAMS object}
\usage{
convert_seurat_to_MAMS(object_list, observation_subsets, dataset_id)
}
\arguments{
\item{object_list}{A named list of Seurat objects to be converted to MAMS format}

\item{observation_subsets}{A vector with same length as object_list indicating the
observation subset name for each Seurat object. One of: full, filtered, threshold, detected,
nonartifact, clean.}
}
\value{
A MAMS object containing all the extracted metadata fields.
}
\description{
Converts Seurat object to a MAMS object
}
\examples{
\dontrun # {mams <- convert_seurat_to_mams("pbmc_seurat")}

}
