% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\name{create_FOM_object}
\alias{create_FOM_object}
\title{Constructor function to create a FOM object}
\usage{
create_FOM_object(
  id = NA_character_,
  dataset_id = NA_character_,
  filepath = NA_character_,
  accessor = NA_character_,
  data_type = NA_character_,
  representation = NA_character_,
  representation_description = NA_character_,
  obs_unit = NA_character_,
  obs_unit_description = NA_character_,
  processing = NA_character_,
  processing_description = NA_character_,
  analyte = NA_character_,
  analyte_description = NA_character_,
  modality = NA_character_,
  obs_subset = NA_character_,
  obs_subset_description = NA_character_,
  feature_subset = NA_character_,
  feature_subset_description = NA_character_,
  record_id = NA_character_,
  parent_id = NA_character_,
  parent_relationship = NA_character_,
  parent_relationship_description = NA_character_,
  oid = NA_character_,
  fid = NA_character_,
  obs = NA_character_,
  fea = NA_character_,
  ong = NA_character_,
  fng = NA_character_
)
}
\arguments{
\item{id}{Denotes the unique id of the matrix, annotation data frame, or graph and should be unique}

\item{dataset_id}{ID of the dataset}

\item{filepath}{Path to the file}

\item{accessor}{Accessor}

\item{data_type}{Explicitly describes the type of data stored in the FOM}

\item{representation}{Preferred representation of the matrix}

\item{representation_description}{More detail about the representation}

\item{obs_unit}{Biological unit of the observations}

\item{obs_unit_description}{Used to describe the obs_unit}

\item{processing}{Used to describe the nature of the data contained within the matrix}

\item{processing_description}{More detail about the nature of the data}

\item{analyte}{Used to describe the biological analytes being quantified in the matrix}

\item{analyte_description}{More details about the analytes}

\item{modality}{Describes the modality of the matrix, may be the same as another field or combination of other fields}

\item{obs_subset}{Describes the subset of observations that are present in the FOM}

\item{obs_subset_description}{More about the subset of observations}

\item{feature_subset}{Describes the subset of features that are present in the FOM}

\item{feature_subset_description}{More about the subset of features}

\item{record_id}{Unique id to denote a combination of entries for record_package_name, record_package_version, record_function_name, and record_function_parameters}

\item{parent_id}{Denotes the id(s) of the parent matrices that were used to produce the matrix}

\item{parent_relationship}{Denotes the type of relationship with the parent matrix or matrices}

\item{parent_relationship_description}{More about the type of relationship with the parent matrix or matrices}

\item{oid}{Character vector or combination of character vectors used to denote the unique ID of each observation}

\item{fid}{Character vector or combination of character vectors used to denote the unique ID of each feature}

\item{obs}{Name of matrices or data frames with the same number of observations as its corresponding FOM}

\item{fea}{Name of matrices or data frames with the same number of features as its corresponding FOM}

\item{ong}{Name of observation neighborhood graph}

\item{fng}{Name of feature neighborhood graph}
}
\value{
a FOM S4 object for further use with other MAMS objects and sub-objects
}
\description{
Create the FOM object
}
