% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\name{create_MAMS_object}
\alias{create_MAMS_object}
\title{Constructor function to create a MAMS object}
\usage{
create_MAMS_object(
  FOM = list(),
  FEA = list(),
  OBS = list(),
  FID = list(),
  OID = list(),
  REC = list(),
  ONG = list(),
  FNG = list()
)
}
\arguments{
\item{FOM}{}

\item{FEA}{}

\item{OBS}{}

\item{FID}{}

\item{OID}{}

\item{REC}{}

\item{FNG}{}
}
\description{
Constructor function to create a MAMS object
}
