% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FEA.R
\docType{class}
\name{FEA-class}
\alias{FEA-class}
\title{class FEA}
\value{
a FEA class for use with MAMS
}
\description{
Stores the feature metadata
}
\details{
Define the FEA (feature) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{dataset_id}}{character.}

\item{\code{filepath}}{character.}

\item{\code{accessor}}{character.}

\item{\code{feature_modality}}{character.}

\item{\code{reference_database}}{character.}

\item{\code{reference_organism}}{character.}

\item{\code{record_id}}{character.}
}}

