% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_FOM.R
\docType{class}
\name{FOM-class}
\alias{FOM-class}
\title{Class to create a FOM object}
\description{
Class to create a FOM object
}
\section{Slots}{

\describe{
\item{\code{id}}{character.}

\item{\code{dataset_id}}{character.}

\item{\code{filepath}}{character.}

\item{\code{accessor}}{character.}

\item{\code{data_type}}{character.}

\item{\code{representation}}{character.}

\item{\code{representation_description}}{character.}

\item{\code{obs_unit}}{character.}

\item{\code{processing}}{character.}

\item{\code{processing_description}}{character.}

\item{\code{analyte}}{character.}

\item{\code{analyte_description}}{character.}

\item{\code{modality}}{character.}

\item{\code{obs_subset}}{character.}

\item{\code{obs_subset_description}}{character.}

\item{\code{feature_subset}}{character.}

\item{\code{feature_subset_description}}{character.}

\item{\code{record_id}}{character.}

\item{\code{parent_id}}{character.}

\item{\code{parent_relationship}}{character.}

\item{\code{parent_relationship_description}}{character.}

\item{\code{oid}}{character.}

\item{\code{fid}}{character.}

\item{\code{obs}}{character.}

\item{\code{fea}}{character.}

\item{\code{ong}}{character.}

\item{\code{fng}}{character.}
}}

