% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_MAMS.R
\docType{class}
\name{MAMS-class}
\alias{MAMS-class}
\title{Class to create MAMS object}
\description{
Class to create MAMS object
}
\section{Slots}{

\describe{
\item{\code{FOM}}{list.}

\item{\code{FEA}}{list.}

\item{\code{OBS}}{list.}

\item{\code{FID}}{list.}

\item{\code{OID}}{list.}

\item{\code{REC}}{list.}

\item{\code{ONG}}{list.}

\item{\code{FNG}}{list.}
}}

