% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_OBS.R
\docType{class}
\name{OBS-class}
\alias{OBS-class}
\title{class OBS}
\value{
the OBS class
}
\description{
Stores the observation data.
}
\details{
Define the OBS (observation) S4 object
}
\section{Slots}{

\describe{
\item{\code{id}}{character}

\item{\code{dataset_id}}{character}

\item{\code{filepath}}{character}

\item{\code{accessor}}{character}

\item{\code{record_id}}{character}
}}

